+++ lib/rubocop/socketry.rb
@@ -7,6 +7,7 @@ require_relative "socketry/version"
 require_relative "socketry/plugin"
 require_relative "socketry/layout/consistent_blank_line_indentation"
 require_relative "socketry/layout/block_delimiter_spacing"
+require_relative "socketry/style/global_exception_variables"
 
 # @namespace
 module RuboCop
@@ -14,6 +15,10 @@ module RuboCop
 	module Socketry
 		# @namespace
 		module Layout
+		end
+		
+		# @namespace
+		module Style
 		end
 	end
 end